added new policy service

+++ Restaurants.Application/Dishes/Commands/DeleteDishesForRestaurant/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,20 +1,33 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishesForRestaurant;
 
 public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRestaurantCommandHandler> logger,
     IRestaurantsRepository restaurantsRepository,
-    IDishesRepository dishesRepository) : IRequestHandler<DeleteDishesForRestaurantCommand>
+    IDishesRepository dishesRepository,
+    IRestaurantsAuthorizationService restaurantsAuthorizationService) : IRequestHandler<DeleteDishesForRestaurantCommand>
 {
     public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
     {
         logger.LogInformation("Removing all dishes from restaurant: {@DishRequest}", request.RestaurantId);
         var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId)
             ?? throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
-        await dishesRepository.DeleteRange(restaurant.Dishes);
+
+        if (!restaurantsAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+            throw new ForbiddenException();
+        try
+        {
+            await dishesRepository.DeleteRange(restaurant.Dishes);
+        }
+        catch (Exception e)
+        {
+            throw new BadRequestException("Error : " + e.Message);
+        }
     }
 }
